Refactor and cleanup.
